% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcec-package.R
\docType{package}
\name{wcec}
\alias{wcec-package}
\alias{wcec}
\title{wcec: ...}
\description{
The \code{wcec} package provides ...
}
\details{
\tabular{ll}{
Package: \tab wcec\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2024-08-07\cr
License: \tab GPL-3\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/swarm-lab/wcec}
  \item \url{https://swarm-lab.github.io/wcec/}
  \item Report bugs at \url{https://github.com/swarm-lab/wcec/issues}
}

}
\author{
Simon Garnier \href{mailto:garnier@njit.edu}{garnier@njit.edu}

Jason Graham \href{mailto:jason.graham@scranton.edu}{jason.graham@scranton.edu}

Maintainer: Simon Garnier \href{mailto:garnier@njit.edu}{garnier@njit.edu}
}
